renamed the main files according to conventions
